requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-filter>=23.4
python-dotenv>=1.0.0
Faker>=24.2.0
django-querycount==0.8.3
django-extensions==3.2.3
django-redis>=5.4.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
